add default actions for mapColors #804
